Various changes to the game
New game.py module created to handle game sessions. It is now possible to set all game session variables in a new Session class. main.py code has been modified accordingly.

A new way to handle enemies has been introduced. Now each planet has a set of enemies related to it, inizialized when a new planet object is instanciated. It is possible to choose the number of enemies linked to each planet, and they will be automatically placed on top of the planet. Support for serialization and de-serialization of enemies related to planets has been added as well.

Collectible items have been added to the game. 3 gems are instanciated when a planet is initialized. The three gems move in circle on top of the planet, and they disappear when the spaceship overlaps them. As a result, game score is incremented by 100.

Spaceship rotation has been included.

Got rid of center_x and center_y variables in the __init()__ method of the GameObject class, since they where not needed explicitly.

Game over text added when the spaceship collides with an enemy.

Serialization and de-serialization are now done with dictionaries.