Lab sessions (#31)
* first commit - snake game complete

* modular snake pygame!

* typo

* add the README.md with challenge explanation

* just moving around stuff

* just moving around stuff

* challenge vs solution organization

* Few modifications

* Few modifications

* Periodicity added to final solution

* Final version

The game over condition on snake touching itself must be implemented.
Initial snake length increased.
Final solution changed accordingly

* Removed some redundand files.

* align with Dario's contribution

* align with Dario's contribution

* working sample - just start thinking about it

* Added incremental level solutions

* a working example of turtle module on the browser

* real time turtle module - just refresh the webpage!

* Added random walls to snake game and pyramid.py exercise

* Adding incremental levels from previous lab-sessions

* Added working menu with pygame_menu

* working pygbag version - without menu

* NOT working pygbag version - with menu

* Changing directory structure

* Campaign mode added

* Added minimal example of endless scrolling

* Added campaign mode

* Adding background and ship images for the space race project

* finally the working version of shifting periodic 2D maps

* All work and no play makes Jack a dull boy!

* Few modifications to the space exploration

* Updating lab-sessions dir structure

* Updated

* Added instructions used throughout lab sessions

* almost working everything

* Working version with classes.

Added classes to handle the map, the planets and the enemies objects.

* Updated version 23/03/2023

* a little refactoring ... and working shots!

* Small modifications to objects position and collision detection

* Copying files to new directory for lab session on serialization

* This a Ronno update

* Adding dario_3 directory

* Various changes to the game

New game.py module created to handle game sessions. It is now possible to set all game session variables in a new Session class. main.py code has been modified accordingly.

A new way to handle enemies has been introduced. Now each planet has a set of enemies related to it, inizialized when a new planet object is instanciated. It is possible to choose the number of enemies linked to each planet, and they will be automatically placed on top of the planet. Support for serialization and de-serialization of enemies related to planets has been added as well.

Collectible items have been added to the game. 3 gems are instanciated when a planet is initialized. The three gems move in circle on top of the planet, and they disappear when the spaceship overlaps them. As a result, game score is incremented by 100.

Spaceship rotation has been included.

Got rid of center_x and center_y variables in the __init()__ method of the GameObject class, since they where not needed explicitly.

Game over text added when the spaceship collides with an enemy.

Serialization and de-serialization are now done with dictionaries.

* Removing old files

* Few modifications to the GameObject/move_around method

* Adding the diamonds

* few last sessions with dario

* first attempt to make the tiber-river google maps client

---------

Co-authored-by: dario-abbondanza <dario.abbondanza@gmail.com>
Co-authored-by: Dario Abbondanza <43699847+dario-abbondanza@users.noreply.github.com>