Add files via upload
Program accepts two inputs, a number to find the approximate square root of "givenNumber" and the guess square root to use in the initialization of Heron's algorithm "guess" 

Calculation begins with the "calculate_sqrt()" function by taking parameter 'x'. The function initializes a counter variable as 'counter' equal to zero to determine the number of iterations, and sets a while loop to calculate the approximate root and add to the counter. The while loop determines if the condition, defined as the absolute value of the Python calculated square root minus the iterated guess is greater than or equal to the error threshold of 10^-12, is true. If true, the loop iterates to add 1 to the variable 'counter' value, sets a variable 'runloop' call the  "HeronsForm()" function to calculate an approximation based on the iteration of the loop using the formula, and then sets variable 'x' to the value of variable 'runloop'. The loop continues to iterate as stated until the condition is verified as false, at which point the function "calculate_sqrt()" will return both variable 'x' as the approximate square root and variable 'counter' as the output.

The function "calculate_sqrt()" is called with both variables 'SquareRoot' and 'NumberOfLoops' to return the square root approximation and loop counter value respectively. The variable 'error_precentage" is defined and calculates the error percentage as the absolute value of variable 'SquareRoot' divided by the Python calculated square root subtracted by 1 with the difference being multiplied by 100 to yield a percentage. Variable 'correct_result' is defined to calculate the Python calculated square root to display to the user.

The final variables are converted to strings as needed and are printed to the user in the console.